# apps/api/docs/swagger.yaml
openapi: 3.0.3
info:
  title: LAFAM API
  version: 0.1.0
  description:
    'LAFAM backend API contract for the single-tenant Pilates and ladies salon booking platform. The contract currently
    exposes Core foundation endpoints and the approved Auth module endpoints. Auth includes public customer sign-up/login/password
    reset, guest sessions, authenticated session/profile management, admin user management, and frontend auth-context bootstrapping.


    **OpenAPI JSON:** [openapi.json](/api/openapi.json)

    '
servers:
  - url: http://localhost:4000/api
    description: Local development API server
...
paths:
...
  /auth/guest-session:
    post:
      summary: Create guest session
      description: Creates an authenticated anonymous guest session with limited permissions.
      operationId: authCreateGuestSession
      tags:
        - Auth - Guest
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGuestSessionRequest'
      responses:
        '201':
          description: Guest session created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthGuestSessionSuccessResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
    delete:
      summary: End guest session
      description: Ends the current active guest session. This does not delete a real customer account.
      operationId: authEndGuestSession
      tags:
        - Auth - Guest
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Guest session ended successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthEndGuestSessionSuccessResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/AuthenticationRequired'
        '403':
          $ref: '#/components/responses/AuthorizationDenied'
  /auth/guest/convert:
    post:
      summary: Convert guest to customer
      description: Converts the current authenticated guest user into a pending customer account and starts email verification.
      operationId: authConvertGuestToCustomer
      tags:
        - Auth - Guest
      security:
        - BearerAuth: []
      parameters:
        - name: x-refresh-token
          in: header
          required: true
          schema:
            type: string
          description: Current guest refresh token required for conversion.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConvertGuestRequest'
      responses:
        '200':
          description: Guest conversion started successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthGuestConversionSuccessResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/AuthenticationRequired'
        '403':
          $ref: '#/components/responses/AuthorizationDenied'
...
components:
  schemas:
    ApiSuccessBase:
      type: object
      required:
        - status
        - message
        - data
        - timestamp_ms
      properties:
        status:
          type: integer
          example: 200
        message:
          type: string
          example: Health check successful
        timestamp_ms:
          type: integer
          format: int64
          example: 1780918065292
...
    ErrorResponse:
      type: object
      required:
        - success
        - error
      properties:
        success:
          type: boolean
          enum:
            - false
        error:
          type: object
          required:
            - code
            - message
            - statusCode
            - path
            - method
            - timestamp
          properties:
            code:
              type: string
              example: AUTHENTICATION_REQUIRED
            message:
              type: string
              example: Authentication is required.
            statusCode:
              type: integer
              example: 401
            path:
              type: string
              example: /api/auth/me
            method:
              type: string
              example: GET
            timestamp:
              type: string
              format: date-time
              example: '2026-06-09T18:20:00.000Z'
            requestId:
              type: string
              example: a3f7bb42-52d6-4b4a-b2fd-a15352b50d11
            details:
              type: object
              additionalProperties: true
              nullable: true
    AuthUserRole:
      type: string
      enum:
        - guest
        - customer
        - admin
        - trainer
        - stylist
        - staff
        - super_admin
    AuthUserStatus:
      type: string
      enum:
        - guest_active
        - pending_email_verification
        - active
        - deactivated
        - deleted
    AuthSessionType:
      type: string
      enum:
        - guest
        - authenticated
        - admin
        - staff
...
    AuthSafeUser:
      type: object
      required:
        - id
        - email
        - phone
        - full_name
        - role
        - status
        - is_guest
        - avatar_path
        - timezone
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
          nullable: true
          example: customer@example.com
        phone:
          type: string
          nullable: true
          example: '+96550000000'
        full_name:
          type: string
          nullable: true
          example: Saqiba Customer
        role:
          $ref: '#/components/schemas/AuthUserRole'
        status:
          $ref: '#/components/schemas/AuthUserStatus'
        is_guest:
          type: boolean
          example: false
        avatar_path:
          type: string
          nullable: true
          example: users/9f35b13e/avatar.jpg
        timezone:
          type: string
          nullable: true
          example: Asia/Kuwait
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    AuthSafeSession:
      type: object
      required:
        - id
        - type
        - device_id
        - device_name
        - ip_address
        - user_agent
        - created_at
        - last_seen_at
        - expires_at
        - revoked_at
        - revoked_reason
        - converted_at
      properties:
        id:
          type: string
          format: uuid
        type:
          $ref: '#/components/schemas/AuthSessionType'
        device_id:
          type: string
          nullable: true
          example: web-chrome
        device_name:
          type: string
          nullable: true
          example: Chrome on Windows
        ip_address:
          type: string
          nullable: true
          example: 203.0.113.10
        user_agent:
          type: string
          nullable: true
          example: Mozilla/5.0
        created_at:
          type: string
          format: date-time
        last_seen_at:
          type: string
          format: date-time
          nullable: true
        expires_at:
          type: string
          format: date-time
          nullable: true
        revoked_at:
          type: string
          format: date-time
          nullable: true
        revoked_reason:
          type: string
          nullable: true
          example: logout
        converted_at:
          type: string
          format: date-time
          nullable: true
...
    CreateGuestSessionRequest:
      type: object
      properties:
        device_id:
          type: string
          maxLength: 128
          nullable: true
          example: web-guest
        device_name:
          type: string
          maxLength: 160
          nullable: true
          example: Guest browser
        captcha_token:
          type: string
          maxLength: 4096
          nullable: true
          example: captcha-token
    ConvertGuestRequest:
      type: object
      required:
        - full_name
        - email
        - password
        - confirm_password
      properties:
        full_name:
          type: string
          maxLength: 120
          example: Saqiba Customer
        email:
          type: string
          format: email
          maxLength: 254
          example: customer@example.com
        phone:
          type: string
          maxLength: 32
          nullable: true
          example: '+96550000000'
        password:
          type: string
          format: password
          minLength: 8
          maxLength: 128
          example: StrongP@ssw0rd
        confirm_password:
          type: string
          format: password
          minLength: 8
          maxLength: 128
          example: StrongP@ssw0rd
        timezone:
          type: string
          maxLength: 64
          nullable: true
          example: Asia/Kuwait
...
    AuthGuestSessionData:
      type: object
      required:
        - access_token
        - refresh_token
        - token_type
        - expires_in
        - user
        - session
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        token_type:
          type: string
          enum:
            - bearer
        expires_in:
          type: integer
        user:
          $ref: '#/components/schemas/AuthSafeUser'
        session:
          $ref: '#/components/schemas/AuthSafeSession'
    AuthGuestConversionData:
      type: object
      required:
        - user
        - email_verification_required
        - guest_converted
      properties:
        user:
          $ref: '#/components/schemas/AuthSafeUser'
        email_verification_required:
          type: boolean
          enum:
            - true
        guest_converted:
          type: boolean
          enum:
            - true
    AuthEndGuestSessionData:
      type: object
      required:
        - guest_session_ended
        - session_id
        - ended_at
      properties:
        guest_session_ended:
          type: boolean
          enum:
            - true
        session_id:
          type: string
          format: uuid
        ended_at:
          type: string
          format: date-time
...
    AuthGuestSessionSuccessResponse:
      allOf:
        - $ref: '#/components/schemas/ApiSuccessBase'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/AuthGuestSessionData'
    AuthGuestConversionSuccessResponse:
      allOf:
        - $ref: '#/components/schemas/ApiSuccessBase'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/AuthGuestConversionData'
    AuthEndGuestSessionSuccessResponse:
      allOf:
        - $ref: '#/components/schemas/ApiSuccessBase'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/AuthEndGuestSessionData'
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Supabase access token accepted by the LAFAM Auth guard.
  responses:
    ValidationError:
      description: Request validation failed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    AuthenticationRequired:
      description: Authentication is required or the token/session is invalid.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    AuthorizationDenied:
      description: The authenticated user is not allowed to perform this action.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...